requirements.txt
Django==1.8.3
django-flat-theme==0.9.5
django-model-utils==2.2
djangorestframework==3.1.3
elasticsearch-dsl==0.0.4
pytest==2.7.2
pytest-django==2.8.0
factory-boy==2.5.2
braintree==3.16.0
mock==1.1.2
Pillow==2.9.0
geopy==1.10.0
whitenoise==2.0.2
